textanalysis: multiple changes to autocomplete
1. move the sugg variable to textanalysis.js instead of the ui part
2. use a callback for the autocomplete widget (it supports three
options: array, string as url for a json returning service, or callback)
3. use a dictionary instead of an array to keep the list of suggestions.

We iterate on the dictionary instead of using a Trie or some other
datastructure in the callback. But the dictionary avoids iteration when
adding a word. Overall a loss in performance, but not noticeable.

Googling / memory suggests a Trie is what we want here, but I'm
postponing that for later. See:

http://ejohn.org/blog/javascript-trie-performance-analysis