rhizicore: set minimal svg size to (1000,1000)
Works around a problem where you load rhizi with a small window size and then
enlarge it, since we don't dynamically change the svg size (and it is
fixed instead of filling a parent div or such) we lose access to a lot
of the screen (i.e. can't drag nodes there, can't see anything there,
etc).