rhizicore: bug fix: draw links below nodes
bug introduced in cd7157955d15802f34dee6ac4578ae249f3e397d
before that we recreated the links and nodes every update, and since all
links were created before nodes the document order was
<link>
<link>
<node>
<node>

which per SVG rendering order [1] leads to nodes being drawn on top of
links.

But after that commit the new links and nodes where created as they were
being edited, so we got
<link>
<node>
<link>
<node>

With this commit all links are created in a single group, getting back
the correct order:

<g>
 <link>
 <link>
</g>
<node>
<node>

nodes are ungrouped but there is no need for that (except that it looks
strangely assymetrical).

[1] http://dev.w3.org/SVG/modules/renderorder/SVGRenderOrder.html