Fix highlight of node and edge
TODO 1:
This is a bit magic. The problem is that I'm not sure exactly if the
nodes/edges given to the click callback are updated correctly when we
just do renames. So the code is a bit protective - it compares names
instead of comparing objects. This might work or not later with multiple
edges. But I really need to fix it so we never have callbacks with stale
data.

TODO 2:
Plus I'm not using classes correctly, instead setting attributes in the
d3 update function (update). This is already done but is a bad smell,
makes it harder to change properties easily. Used for stroke-width here
and for nodes, for color.