add a copy to url button, handle the resulting url
This allows us to send small graphs (up to maximal length of query
string - what is that anyway?). Just press the copy button, ctrl-c, and
paste to irc, and other side should click it.

The real deal, issue #65, needs to actually go to the server and load
the graph, which means the server needs to have a document concept.