textanalysis.ui: fix for lingering #102 issue
After using up/down keys to select one of the entries you had an extra
enter required to "complete" the sentence.

Reproduce:
ctrl-alt-i (focus on 'Type here to create a network', #textanalysis)

The first enter is the selection, the second is ignored (the bug fixed
by this commit), the third is handled by the #textanalysis keypress
trigger.

The fix was simple, not sure why it works. Seems there is a
keyup/keypress interaction. The keyup was on the document element,
replaced by a keypress on the same element.