Fixes #295 adding a link next to node names with urls
Note about implementation: we currently update all the nodes for
multiple things, we should not do that, we have id's on each node, we
can do less work. Another option is using the visualdom and letting it
do the diffs. Actually we should test first. Which brings me to the
missing functionality of loading a large csv.