client/graph_view: apply zoom of each node individually
Fixes #308

Hack for friday. Hack being that the model Node gets a 'zoom_obj'
property. This belongs on a new NodeView class that will contain a
reference to the node and the zoom object.

Also I'm worried about performance of this (well, just a little), have not tested with
something large yet.