client/textanalysis.ui: set typeselection position at node, fixes #261
Note that the selection doesn't move with the node, so it is possible to
create a discrepancy, this can happen either when:
1. dragging or zooming the graph
2. causing the bubble animation to playout

The selection text will not be next to the node in both cases until Tab
is pressed again.

Fixing this right now seems too much work, will revisit if this is
really bothersome. Would require either timers or another connection, or
different implementation as a foreign object under the svg node (and so
tied by the layout, no need to change).