client: fix scrolling of window on focus in feedback button (Fixes #366)
This is actually a sort of workaround, since now you cannot use tab to
focus on the feedback button, you have to click it with the mouse.

Since most people (everybody?) would do that anyhow, it's good enough.

An alternative is to change the 'botton: -3px' to anythin non-negative.

I have not managed to capture or prevent the scrolling behaviour. What
happens is that chrome tries to move the element into focus even though
that is impossible since it is with a fixed position to an absolutely
positioned element (body).