client: convert svg icons to png
while deploying a docker container both firefox and chrome refuse to
load same path svg (reporting 'image failed to load' but showing 200 on
the client, suggesting it is a behaviour implemnted after the HTTP
interaction is completed successfully), while they accept png. This is a
workaround. We already have other icons as png's, and we do not support
change in icon size, suggesting no down side to this right now. The svg
sources are left, and are actually still deployed (minor issue).