adding scripts/build-debian
It is specialized in that it uses a systemd-nspawn command to launch the
building command and clone the repository. It runs pretty fasts (10
seconds) on my machine (Fedora 23), the container is Debian 7 with some
additional packages installed - basically all our requirements, but as
deb packages (version checking hasn't been done), plus debian build
packages.