client: different sized nodes per node type
Usage:

In domain_types add a key 'radius' to the node type you want, with
value an array whose first parameter is the attribute. The rest of the
array is currently ignored. The attribute will be parsed as a float, if
the parsing fails it will be ignored. If there are other nodes which
have the radius defined, the undefined nodes will have an undefined
radius.

Example:

'project': {
 'radius': ['total-hours']
}