rhizi: broadcast attr diffs to all relevant documents
Since a Topo_Diff only affects one document, the only remaining
operation is Attr_Diff, and it can affect many documents. Introduce
DBO_find_rzdocs_touching to return all documents touching on a given set
of node ids and link ids. The clients to broadcast are the sum of the
clients listening to those documents.